bug fix
fixed a bug where the new TRANS opcode family didn't actually have its handler set in exetables.cpp.
this was causing it to be flagged as an unknown opcode during execution.

found out there was a bug in the 64-bit multiply code because of a bug in the biggerints submodule.
fixed that and merged the updates in.

+++ src/Instructions.cpp
@@ -3293,7 +3293,7 @@ namespace CSX64
             {
                 if ((mask & 1) != 0)
                 {
-                    if (!SetMemRaw(m, Size(elem_sizecode), ZMMRegisters[reg].uint((int)elem_sizecode, i))) return false;
+                    if (!SetMemRaw(m, Size(elem_sizecode), ZMMRegisters[reg].uint(elem_sizecode, i))) return false;
                 }
                 else if (zmask && !SetMemRaw(m, Size(elem_sizecode), 0)) return false;
             }
